requirements.txt
gunicorn>=21.2.0

# Core
Django==4.2.10
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
typing_extensions==4.15.0

# REST / Auth
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-oauth-toolkit==3.1.0
oauthlib==3.3.1
jwcrypto==1.5.6

# AWS / Storage (Python 3.12 SAFE)
boto3>=1.34.0
botocore>=1.34.0
django-storages>=1.14.4
urllib3>=1.26.18,<2.1
requests>=2.31.0

# Background / Tasks
celery>=5.3.0
schedule==1.2.2
psutil==7.1.3

# Security / Crypto
cryptography>=46.0.0
cffi>=2.0.0
pycparser==2.23
certifi>=2024.2.2
charset-normalizer>=3.3.0
idna==3.11
six==1.17.0

# Screen / Monitor (mac compatible)
mss==10.1.0
MouseInfo==0.1.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0

# Build / Packaging
pyinstaller==6.17.0
pyinstaller-hooks-contrib>=2024.0
altgraph==0.17.5
packaging==25.0
setuptools>=69.0.0

# PDF / Reports
reportlab==4.4.7

# Payments
razorpay==2.0.0